modify docker.sh add distrobox.sh

+++ docker/install-ros-noetic-in-docker.sh
@@ -2,14 +2,37 @@
 
 echo ""
 echo "#######################################################################################################################"
-echo ">>> update apt repo and upgrade, install git vim nano neovim and config $TERM ,then install keyboard-configuration"
+echo ">>> update apt repo and upgrade, install git vim nano neovim and config TERM ,then install keyboard-configuration"
 echo ""
 echo "#######################################################################################################################"
 sudo apt update
 sudo apt upgrade
-sudo apt install -f git vim nano neovim
-export TERM=xterm
-sudo apt install -f keyboard-configuration
+read -p "Do you want to install git vim nano neovim? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y"]; then
+	sudo apt install -y git vim nano neovim
+fi
+read -p "Do you want to export TERM=xterm and install keyboard-configuration? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	export TERM=xterm
+	sudo apt install -y keyboard-configuration
+fi
+
+echo ""
+echo "#######################################################################################################################"
+echo ">>> install lsb-release wget curle from apt"
+echo ""
+echo "#######################################################################################################################"
+
+read -p "Do you want to install lsb-release? [y/n](in docker ubuntu20.04 it doesn't installed)" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	sudo apt install -y lsb-release wget curl
+fi
 
 echo ""
 echo "#######################################################################################################################"
@@ -17,8 +40,19 @@ echo ">>> install vscode from offcial"
 echo ""
 echo "#######################################################################################################################"
 
-wget https://az764295.vo.msecnd.net/stable/abd2f3db4bdb28f9e95536dfa84d8479f1eb312d/code_1.82.2-1694671812_amd64.deb
-sudo apt install -f ./code_1.82.2-1694671812_amd64.deb
+read -p "Do you want to install vscode? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	wget https://az764295.vo.msecnd.net/stable/abd2f3db4bdb28f9e95536dfa84d8479f1eb312d/code_1.82.2-1694671812_amd64.deb
+	sudo apt install -y ./code_1.82.2-1694671812_amd64.deb
+fi
+read -p "Do you want to remove vscode.deb? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	sudo rm ./code_1.82.2-1694671812_amd64.deb
+fi
 
 echo ""
 echo "#######################################################################################################################"
@@ -26,7 +60,12 @@ echo ">>> install libpcl-dev libopencv-dev libeigen3-dev libboost-all-dev from a
 echo ""
 echo "#######################################################################################################################"
 
-sudo apt install -f libpcl-dev libopencv-dev libeigen3-dev libboost-all-dev
+read -p "Do you want to install those packages? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	sudo apt install -y libpcl-dev libopencv-dev libeigen3-dev libboost-all-dev
+fi
 
 echo ""
 echo "#######################################################################################################################"
@@ -34,9 +73,43 @@ echo ">>> install lazygit from github"
 echo ""
 echo "#######################################################################################################################"
 
-LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
-curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
-tar xf lazygit.tar.gz lazygit
-sudo install lazygit /usr/local/bin
+read -p "Do you want to install lazygit? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
+	curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
+	tar xf lazygit.tar.gz lazygit
+	sudo install lazygit /usr/local/bin
+fi
+read -p "Do you want to remove lazygit cache? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	sudo rm lazygit lazygit.tar.gz
+fi
 
-curl https://raw.githubusercontent.com/qboticslabs/ros_install_noetic/master/ros_install_noetic.sh | bash
+echo ""
+echo "#######################################################################################################################"
+echo ">>> install ros noetic use the install script from github"
+echo ""
+echo "#######################################################################################################################"
+read -p "Do you want to install ros noetic use the install script? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	curl https://raw.githubusercontent.com/qboticslabs/ros_install_noetic/master/ros_install_noetic.sh | bash
+fi
+
+
+echo ""
+echo "#######################################################################################################################"
+echo ">>> source ~/.bashrc"
+echo ""
+echo "#######################################################################################################################"
+read -p "Do you want to exec it? [y/n]" answer
+#Default answer is y
+answer=${answer:-y}
+if [ $answer == "y" ]; then
+	source ~/.bashrc
+fi
